Set "suspended" when a surface is minimized
This seems for an SDL XWayland client to restore fullscreen after
unminimize, it needs to see the `_NET_WM_STATE_FOCUSED` state get set
and unset.

This is a somewhat awkward way to do this. I'm not sure if the Wayland
`suspended` flag should be treated as equivalent to `_NET_WM_STATE_FOCUSED`?
If it's meant to be, the way it's defined in the protocol doesn't quite seem
right.

https://github.com/pop-os/cosmic-comp/issues/1510